requirements.txt
openshift
python-dotenv
requests